GEOMESA-1866 Refactoring Arrow scans for shared use
