+++ geomesa-index-api/pom.xml
@@ -37,6 +37,13 @@
             <artifactId>mango-core</artifactId>
         </dependency>
 
+        <!-- pull in arrow for aggregating scan code, but don't make it a transitive dep -->
+        <dependency>
+            <groupId>org.locationtech.geomesa</groupId>
+            <artifactId>geomesa-arrow-gt_${scala.binary.version}</artifactId>
+            <scope>provided</scope>
+        </dependency>
+
         <!-- test dependencies -->
         <dependency>
             <groupId>org.specs2</groupId>
